+++ src/Import/FileImport.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Taskforce\Import;
+
+class FileImport extends BasicImport {}
+
+?>
